+++ src/app/services/archive.service.ts
@@ -34,9 +34,8 @@ import {
   seriesProgramsMethod,
   translationMethod,
   searchMethod,
-  networkKey,
-  yesKey,
-  noKey
+  errorTextKey,
+  networkRequestFailedText
 } from '../helpers/constants';
 
 @Injectable({
@@ -413,18 +412,26 @@ export class ArchiveService {
     let xhttp = new XMLHttpRequest();
     xhttp.onload = (e) => {
       //console.log('Response: ', xhttp.responseText);
-      this.commonService.cacheValue(networkKey, yesKey);
 
       cb(xhttp.responseText);
     };
 
     xhttp.onerror = (e) => {
       //console.log('Network request failed: ', e);
-      this.commonService.cacheValue(networkKey, noKey);
+      this.commonService.cacheValue(errorTextKey, networkRequestFailedText);
 
       cb(null);
     };
 
+    xhttp.ontimeout = (e) => {
+      //console.log('Network request timeout: ', e);
+      xhttp.abort();
+
+      this.commonService.cacheValue(errorTextKey, networkRequestFailedText);
+      cb(null);
+    }
+
+    xhttp.timeout = 30000;
     xhttp.open('GET', url, true);
     xhttp.send();
   }